Extract configuration extention
